fix(core): fix vitest test runner options for angular in cnw (#33921)
## Current Behavior

Creating a new Angular workspace with Vitest results in no test setup
being generated. This happens because the `vitest` option is no longer
available, and instead, there are two options: `vitest-angular` and
`vitest-analog`.

## Expected Behavior

Creating a new Angular workspace should prompt for Vitest with Angular
or Vitest with Analog to set up the tests.

---------

Co-authored-by: nx-cloud[bot] <71083854+nx-cloud[bot]@users.noreply.github.com>
Co-authored-by: leosvelperez <leosvelperez@users.noreply.github.com>
Co-authored-by: Jack Hsu <jack.hsu@gmail.com>